Time: 0 ms (100.00%), Space: 41.8 MB (5.14%) - LeetHub

+++ 62-unique-pathsa-idbrief-btn-62-href-target-blank-classbutton-4-stylefont-weight-bold-margin-left-10px-a/62-unique-pathsa-idbrief-btn-62-href-target-blank-classbutton-4-stylefont-weight-bold-margin-left-10px-a.java
@@ -1,6 +1,6 @@
 class Solution {
     public int uniquePaths(int m, int n) {
-        long[][] dp = new long[m][n];
+        int[][] dp = new int[m][n];
         
         dp[0][0] = 1;
         
@@ -13,6 +13,6 @@ class Solution {
             }
         }
         
-        return (int) dp[m - 1][n - 1];
+        return dp[m - 1][n - 1];
     }
 }
